Compute sound speed on faces correctly

+++ plot.py
@@ -3,7 +3,7 @@ import math
 import numpy as np
 import matplotlib.pyplot as plt
 
-primitive = np.fromfile(sys.argv[1])
+primitive = np.fromfile(sys.argv[1], dtype=np.float64)
 n = math.isqrt(len(primitive) // 3)
 primitive = primitive.reshape([n, n, 3])
 # plt.plot(primitive[n//2,:,0])
